requirements.txt
discord.py
arxiv
openai
transformers
google-generativeai
polars